Remove unused functions from functions.py
The parse_xxx functions were used in version 1 to skip past any text
that did not contain characters from the desired number system. This
was only necessary because the bot would prepend the output text with
"Encoded ASCII binary:" or "Decoded ASCII text:". This is removed in
version 2, and thus the parse functions are no longer necessary. The
comment summoning ASCIICodecBot should provide enough context for the
bot's response.

Additionally, several comments were added in functions.py clarifying
the purpose of the functions in use.