Clip lines in README.md and constants.py
This fix should resolve some legibility issues. Namely:

1. The README file had previously had each paragraph written on a single
line. While this is not noticeable on GitHub or another MarkDown viewer,
this impeded the legibility of the file on Vim or other text editors. I
adhered to the standard 72 character limit for commit messages.

2. The help message provided by the bot through `u/asciicodecbot help`
had the same issue. While not typically an issue on desktop Reddit
clients, the message would often exceed the width of the screen on
mobile clients and require vertical scrolling. This should resolve that.