Add simple exception logging
A simple logging functionality was added to document any exceptions that
may occur while I am not by the computer to monitor the stack trace
output onscreen. Each entry includes what should be enough information
to get the gist of what caused the error, including an identifier for
the comment that summoned ASCIICodecBot.

Some unused commented code was also removed for clarity.