Add error message for missing command error
The bot will not reply if there is only a username mention with no
command. This might be confusing for users, since it may imply that the
bot did not see the comment. This, in conjunction with the basic usage
intructions, will remind the user of their failure to properly use the
bot.