Add manually thrown exception
The command is not documented in the help or usage messages, as it is
used only for debug purposes. The "throw" command triggers an
AsserionError and thus writes an error log entry. The entry itself
should not differ from any other exception that may be thrown during
normal operation, so I added it to give myself the option to trigger log
entries manually when nobody is posting comments that break the bot.