Change comment fetching to continuous stream
This upgrade is backend-only, and should be imperceptible from the
user's perspective.

The script would previously fetch any new comments from the inbox at a
regular interval. There was nothing severely wrong with this strategy
from a user standpoint, other than the delay from comment to reply would
sometimes be as much as 30 seconds. However, this had the inherent
effect of cluttering up the console with unnecessry messages every 30
seconds letting me know that 0 replies were sent. The scriptnow takes
advantage of the stream_generator function in PRAW to have comments
yielded as they arrive in the inbox, and idle while awaiting them. This
drastically cuts down on spam in the console and can be useful for
debugging if I just want to check the script's output instead of the
logs since the useful information has already been covered by more
useless info.

Instead of printing how many replies were sent in each iteration of
fetching inbox comments, the script now displays a running sum of all
replies sent since the entrypoint.

This change also means that the interval command line argument is no
longer used, as it will have no effect on the program execution.