Fix base64 decoding bug
This bug was not present in every string decoded. Rather, it would only
present itself when certain characters appeared in the base64 data
block. I tracked down the error to several `if` blocks that should have
been `elif` blocks, in the section of code responsible for translating
ASCII characters into base64 integers.