Create functions.c and write first function
The first function is not finished, but I am not going to sleep without
committing my changes first. I need to figure out how to deal with
memory allocation since there will be no main function calling the
encode/decode functions to act as a wrapper for dynamic memory
allocation.