Add manual memory freeing function
Since the decodeBin () function will return a pointer to a wchar_t
array, I need a way for that memory to remain allocated while the python
script can read it and use it. Adding a dedicated freeing function
should bring me one step closer to being able to manage memory through
the python script. The idea is that the script will call the freeing
function on the resultant pointer after it does whatever is necessary
with the string.