Integrate hex and bin C functions
This was a bit of a detailed modification. Four functions have been
rewritten in C in functions.c as of now, and they have been linked into
the python script:

- encodeBin ()
- decodeBin ()
- encodeHex ()
- decodeHex ()

This is in an effort to optimize the performance of these functions in a
way that takes advantage of the low-level nature of C, as well as the
way it deals with strings and characters much more flexibly than Python.

Ideally, all of the functions defined in functions.py will end up being
wrappers for the functions defined in functions.c. I have chosen to make
them wrappers since it would not make sense to clutter up the main
script with code that manually manages pointers and memory. This will
also make it simpler to integrate functions.py into any other projects
if desired, as it will add a layer of abstraction from memory
management.

One final function was added, which is freewchar () from functions.c.
Its purpose is to free the memory dynamically allocated for the string
returned by the conversion functions. It is crucial that it is called
after any call to a function from functions.c, which is another reason
why the functions in functions.py are now wrappers. They reduce the
possibility of calling a C function and forgetting to free the memory
afterward.