Optimize decodeBin ()
decodeBin () calls strlen () three times for each iteration of the while
loop. Since the length of sBinStr will not change at all throughout the
loop, the result of strlen (sBinStr) can safely be stored in a variable
for faster access. This should significantly reduce the amount of time
decodeBin takes to execute.