Allocate enough memory for encoding functions
Due to the nature of encoding plaintext into a series of numbers, the
output string of any of the three basic encoding functions will always
be longer than the input string. Under normal circumstances this would
be fine; however, Reddit does have a hard limit on the length of
comments, which is 10,000 characters.

The changes I've implemented should solve the following problems:

1. The previous value of MAX_COMMENT_LENGTH was 40,000, because I based
   the value off of a google search. I have played around with Reddit
and seen that the limit is actually 10,000. This should theoretically
save some memory during runtime. More importantly, the functions will no
longer spit out strings that may exceed the limit and cause an API
exception during runtime. There is now a hard-coded limit on the length
of the output string which complies with Reddit's limit by truncating
the output string at exactly 10,000 characters.

2. Previously, the encoding functions would occasionally cause a
   segfault, due to them trying to write to an array that was too short
and thus going out of bounds. This was solved by allocating just enough
memory to perform the encoding function on the longest acceptable string
defined by MAX_COMMENT_LENGTH.