Use less memory for encodeBin () and encodeHex ()
Since encodeBin () and encodeHex () both output a constant number of
characters for every input character, it is easy to just truncate the
input string before doing any operations, thus avoiding the use of extra
memory. The number of maximum input characters is defined as the floor
of (MAX_COMMENT_LENGTH / characters of output per input).

The same shortcut cannot be taken with encodeDec (), because the number
of output characters per input character varies with the value of each
character.