Rewrite decodeHex to use char* for input/output
The decodeHex() function in functions.c no longer uses wchar_t* as its
argument and return types. This is part of my effort to make the
functions easier to use and more predictable; i.e., not reliant on
wcstombs() to convert wide character input to byte strings for
processing.

decodeHex() now makes a copy of the input string, since the characters
of that string need to be modified during the capitalization process. In
order to avoid modifying the string pointed to by hexStr, a copy is made
in dynamically allocated memory and freed before the function exits.

A change was also made to both the decodeBin() and decodeHex()
functions. For consistency, the index variable now only declared, not
initialized. It is to be set immediately before use, such as before a
while loop or at the beginning of a for loop. This eliminates any
ambiguity about the state of the variable at the beginning of any loop,
by explicitly assigning it a value before the loop.