Rewrite encodeBin() and encodeHex() in functions.c
encodeBin() and encodeHex() are rewritten to use const char* as their
argument and return types. Their respective wrapper functions in
functions.py are also modified to reflect these changes.

decodeBin():
The type of num has been changed from char to unsigned char. This is to
avoid any problems when doing 8-bit arithmetic, since using signed char
to represent numbers greater than 127 yields unexpected arithmetic
errors.
The type of bitNo was changed to char, since it will never exceed 8
during operation

encodeBin():
Variable i was renamed to num, just to be a little more explanatory with
the meaning of the variable.
In the second for loop, the left-shift assignment operation on i was
removed. This code was left in from when the python code was ported to
C, and it is no longer necessary since numbers greater than 0xFF are
naturally truncated when stored in a char.

decodeHex():
The type of num was changed to unsigned char for the same reason as in
encodeBin().

encodeHex():
Variables i and j were changed from char to unsigned char, for the same
reason as in encodeBin() and decodeHex().
In the for loop, the division assignment operator was changed to a
right-shift assignment operator, because that better reflects what I am
trying to accomplish with the code.
Previously, in the for loop, I had an else clause at the end of the
first if-else ladder. This is not actually necessary, since there is no
way for j to be anything greater than or equal to 16 after having been
bitwise-ANDed with 0xF.
In the same loop, the subtraction-assignment operator was changed to a
bitwise-AND assignment operator, because that also better reflects what
I am trying to accomplish with the code.
Where hexString is appended with a null terminator, I changed the
representation from 0 to '\0' to better reflect the intention of the
line of code.