Rewrite encodeDec() and decodeDec() in functions.c
Both decimal functions have been rewritten to use const char* as their
return type.

decodeDec():
buf was changed to u_int32_t, to make the best use of the int type,
since all integers read from decStr will be positive.
index is no longer initialized when declared. It is assigned a value
whenever it is to be used, such as at the beginning of a loop. This
is so that the code more clearly shows the state of the variable before
a loop.
Previously, decodeDec would return an empty string if decStr began with
a space, because a null byte would be written to byte 0 of the output
string. A loop was added to skip past any leading spaces and stop at the
first decimal character, thereby avoiding returning an empty string
unless decStr explicitly begins with the number 0 expressed in text.

encodeDec():
Argument type was changed to const unsigned char* instead of const
char*. With exclusively-ASCII characters, this wouldn't be a problem,
since ASCII characters do not exceed 127. However, in order to be
UTF-8-compatible, the function must treat the input string as unsigned,
due to values above 127 being interpreted as negative values. This is
not an issue in the other functions; however, since this function uses
characters from the input string without implicitly casting to unsigned
char first, we must treat the entire string as unsigned char.
A longer comment was added to the declaration of decimalString in order
to clarify why its size is defined as MAX_COMMENT_LENGTH * 4 + 1.
The while loop was changed to a for loop for clarity.
There was an if statement checking whether the character in question was
between 0 and 255 inclusive. This check is also a remnant of from when
the code was ported to Python; it is not necessary since an unsigned
char is always positive and always truncated to 8 bits.