Refactor encodeBase64() in functions.c
Argument and return types were changed to variants of char*, to accept
UTF-8 input.
The formula for how much memory to allocate for the output string has
changed. Now, exactly enough memory will be allocated to contain the
entire output string, without truncation.
The function no longer checks for `invalid characters' in the input
string. This was previously done in order to avoid non-ASCII input
producing useless invalid output. However, since the input can
technically be any non-zero byte, this check has been removed. This was
also done because there are valid UTF-8 bytes that are technically not
valid ASCII bytes.
The if statement previously on line 460 was removed, in order to make
the code more readable. Now, that entire while loop is dedicated to
converting all the input bytes that fit into groups of 3. Any leftover
input bytes will be handled by the second while loop.