Return null pointer on error
Now instead of returning a special string signifying that the input is
invalid, the decode functions will return a null pointer, since that
tends to be the general consensus for signaling an error in a function
that returns a pointer. The python wrapper functions for the decode
functions have been modified to work with this change.